fix : apply feedback in pull request
