--- components/header/header.css
@@ -1,49 +1,0 @@
-:host {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  z-index: 1000; /* Ensure the header appears above other content */
-  padding: 20px;
-  border-bottom: 0.1rem var(--border);
-}
-
-header {
-  background-color: rgba(255, 255, 255, 0.8); /* Background color with 80% opacity */
-  width: 80%;
-  margin: 0 auto; /* Center the content horizontally */
-}
-
-.container {
-  display: flex;
-  justify-content: space-between; /* Space between logo and buttons */
-  align-items: center;
-  margin: 0 auto;
-}
-
-header img {
-  width: 4rem;
-  height: 2rem;
-}
-
-header span {
-  font-size: 2rem;
-  margin-left: 1rem;
-}
-
-header img,
-span {
-  vertical-align: middle;
-}
-
-.buttons-container {
-  display: flex;
-}
-
-a {
-  font-size: 1.6rem;
-  font-weight: medium;
-  margin-right: 4rem;
-  color: black;
-  text-decoration: none;
-}
