+++ main.js
@@ -1,6 +1,7 @@
 import "./components/divider/divider.js";
 import "./components/footer-link/footer-link.js";
+import "./components/header.js"
 import "./components/hero/hero.js";
 import "./components/link-button/link-button.js";
 import "./components/seo-meta-tag/seo-meta-tag.js";
-import "./components/header/header.js"
+
